backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
psycopg2-binary==2.9.9
alembic==1.13.2
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pydantic-settings==2.4.0
email-validator==2.1.0
httpx==0.27.2
stripe==10.12.0
Jinja2==3.1.4
pydyf==0.10.0
WeasyPrint==60.2
